requirements.txt
python-dotenv
openai
python-telegram-bot
requests
urllib3<2.0
requests-toolbelt